pyproject.toml
[tool.poetry]
name = "samane-golestan"
version = "0.1.0"
description = "Samane Golestan, an LMS for universities"
authors = ["Zahra Sajadi", "Sepehr Azimi", "Mohammad Ahmadi", "Sahar Haghighat", "Maryam Naseri", "Ali Mohammadzadeh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.1"
markdown = "^3.6"
pygments = "^2.17.2"
django-filter = "^24.1"
django-guardian = "^2.4.0"
pillow = "^10.3.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
pre-commit = "^3.6.2"
poetry-plugin-export = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
